// frontend/src/app/api/model/readSectorGroupVersion.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Status } from './status';


export interface ReadSectorGroupVersion { 
    /**
     * Object creation date
     */
    readonly creationDate?: string;
    /**
     * User creator
     */
    readonly creator?: string;
    /**
     * Last edition date
     */
    readonly editionDate?: string;
    /**
     * User editor
     */
    readonly editor?: string;
    /**
     * This ID helps identify versions of a traffic point element in the use case front end and/or update. This ID can be deleted if the version is no longer present. Do not use this ID to map your object to a traffic point element. To do this, use the sloid in combination with the data range (valid from/valid until). 
     */
    readonly id?: number;
    /**
     * Unique code for sector that is used in customer information. https://transportdatamanagement.ch/standards/
     */
    readonly sloid?: string;
    /**
     * Unique code for traffic point element (TPE) that is used in customer information. By means of this ID, the connection between stops and bus / station stop area or boarding area can be established. The structure is described in the “Swiss Location ID” specification, chapter 4.2. The document is available here. https://transportdatamanagement.ch/standards/ 
     */
    trafficPointSloid: string;
    validFrom: Date;
    validTo: Date;
    /**
     * Designation used in the customer information systems.
     */
    designation: string;
    /**
     * Length of a Sector in m
     */
    length?: number;
    /**
     * Optimistic locking version - instead of ETag HTTP Header (see RFC7232:Section 2.3)
     */
    etagVersion?: number;
    status?: Status;
}
export namespace ReadSectorGroupVersion {
}